Readme Updates
This patch updates the main readme file as well as renames
`PageAllVolumes` to `PageVolumes`.
